Add single point of failure mode and error detect
TODO(): Integrate with proper command parsing,
when this is merged in:
https://github.com/quanta-bmc/read-margin-temp/commit/58a81bf63e4ebb026b2272bbadc8c707b5c25a5b

TODO(): If merged as-is, this will change default behavior,
so it would be safer to have SPOF mode default to disabled,
instead of defaulting to enabled.

TODO(): Test the file-based I/O, make sure it did not break,
as I have only tested it with D-Bus I/O.

Cleaning up std::isnan() usage to std::isfinite(),
which catches more weird floating-point things like +Inf.

Added spofEnabled global option,
enabling it by default, although this should be changed,
to avoid breaking upstream users unexpectedly.

Better istream-to-double reader that catches errors,
although this is not needed for our usage,
adding it for completeness and upstream users.

Flagging errors that happen during sensor reading,
to catch failure-of-any sensor, not just failure-of-all, and
converting failure-of-any to failure-of-all if spofEnabled.

Signed-off-by: Josh Lehan <krellan@google.com>