Caching services and values for fast performance
Adding service cache, to avoid repeated calls to getService(), as the
returned information is unlikely to change. Any error resulting from
calling the service will clear this cache, in case the service name
actually did change.

Adding value cache, to cache the last good value received, from each
sensor. Using D-Bus message notifications, to become aware of changes
to a sensor, instead of polling the value every time through the main
loop. This avoids repeated polling for sensors which are maintaining
a stable temperature. Only good readings are cached, any invalid
reading also clears this cache.

An optional timeout feature is also added, in case message
notifications are not working well on your system. After the timeout,
the value cache for this sensor will be cleared, even if the reading
was good, so another poll will then be made.

A separate cache debugging flag is also added, so it can be enabled
independently of the main margin zone computation debugging flag.

Alarm notifications are also cached, similar to values. When an alarm
indicates a bad condition, cache usage is intentionally disabled, and
the value for this sensor is disregarded, until condition clears up.

Tested: Works nicely locally for me, sensor loop now is much faster.

Signed-off-by: Josh Lehan <krellan@google.com>